[Ref] Move CustomAttribute related async actions to AsyncCustomizationActions.ts
